requirements.txt
kafka-python
fastapi[all]
# faust
faust-streaming
configloader[all]
python-dotenv
Faker